new set up with classes saved in module
